Refactor components structure (#49)
* refactor: move components out of app folder

* refactor: remove unused imports
