+++ Frontend/app/admin/page.tsx
@@ -1,15 +1,15 @@
 "use client"
 
 import { useState, useEffect } from "react"
-import Header from "../components/Header"
-import Footer from "../components/Footer"
+import Header from "@/components/Header"
+import Footer from "@/components/Footer"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
 import Link from "next/link"
 import { PlusCircle, UserPlus, UserCog } from "lucide-react"
-import CreateProposalModal from "../components/CreateProposalModal"
-import CreateUserModal from "../components/CreateUserModal"
-import ModifyUserModal from "../components/ModifyUserModal"
+import CreateProposalModal from "@/components/CreateProposalModal"
+import CreateUserModal from "@/components/CreateUserModal"
+import ModifyUserModal from "@/components/ModifyUserModal"
 import { useContractCustom } from "@/hooks/use-contract"
 import { ProposalPublic } from "@/interfaces/Proposal"
 import { useWallet } from "@/hooks/use-wallet";
