+++ Frontend/app/results/page.tsx
@@ -1,15 +1,13 @@
 "use client";
 
-import Footer from "../components/Footer";
+import Footer from "@/components/Footer";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Progress } from "@/components/ui/progress";
 import { Button } from "@/components/ui/button";
 import Link from "next/link";
 import { useContractCustom } from "@/hooks/use-contract";
 import { useEffect, useState } from "react";
 import { ProposalPublic } from "@/interfaces/Proposal";
-import { useAccount } from "@starknet-react/core";
-import Header from "../components/Header";
+import Header from "@/components/Header";
 import { useWallet } from "@/hooks/use-wallet";
 import { loginStatus } from "@/interfaces/Login";
 
@@ -55,9 +53,8 @@ export default function ResultsPage() {
                         className="h-2.5 rounded-full"
                         style={{
                           backgroundColor: "#f7cf1d",
-                          width: `${
-                            (option.count / voting.total_voters) * 100
-                          }%`,
+                          width: `${(option.count / voting.total_voters) * 100
+                            }%`,
                         }}
                       ></div>
                     </div>
