+++ Frontend/components/CandidateList.tsx
