+++ Frontend/components/Footer.tsx
